refactor: components 디렉토리 개선
