+++ frontend/src/components/index.ts
@@ -1,8 +1,8 @@
 export { default as NavbarSection } from "./navbar/NavbarSection";
-export { default as ProfileBox } from "./ProfileBox";
+export { default as ProfileBox } from "./main/ProfileBox";
 export { default as TopBackground } from "./common/TopBackground";
 export { default as TopBar } from "./common/TopBar";
-export { default as RegisterPageTitle } from "./register/RegisterPageTitle";
-export { default as ThemeButton } from "./ThemeButton";
-export { default as RegisterInputLabel } from "./register/RegisterInputLabel";
+export { default as Title } from "./register/Title";
+export { default as ThemeButton } from "./common/ThemeButton";
+export { default as InputLabel } from "./register/InputLabel";
 export { default as SubmitInput } from "./register/SubmitInput";
