+++ frontend/src/components/main/ProfileBox.tsx
