+++ frontend/src/components/register/InputLabel.tsx
